requirements.txt
pyyaml~=6.0
pycryptodome
cryptography~=37.0.1
PyQt5~=5.15.7
ntplib~=0.4.0
requests~=2.28.1